requirements.txt
mmcv-full==1.3.14
mmdet==2.14.0
mmsegmentation==0.14.1
scikit-image==0.18.3
lyft_dataset_sdk
networkx==2.2
numba==0.48.0
numpy==1.21.0
nuscenes-devkit
plyfile
tensorboard
trimesh==2.35.39
torchmetrics
fvcore
einops
# torch_tb_profiler
# accelerate
# torchsummary
# https://mmcv.readthedocs.io/en/latest/get_started/installation.html#install-mmcv-full